+++ db/migrate/20160607144122_add_authentication_token_to_adverse_events.rb
@@ -1,0 +1,6 @@
+class AddAuthenticationTokenToAdverseEvents < ActiveRecord::Migration
+  def change
+    add_column :adverse_events, :authentication_token, :string
+    add_index  :adverse_events, :authentication_token, unique: true
+  end
+end