+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 0.38.0
 
 ### Enhancements
+- **Adverse Event Changes**
+  - Project and site editors can now generate links to send adverse event
+    summaries to medical monitors
 - **Design Changes**
   - Improved padding between section heading images and surrounding elements
   - Required design options now have an indicator on the design reorder page